Added calibration factor getting, fixed getting the info from the sliders, added a script that should help with cirlce finding
Fixed the scope problems getting the info from the ui objects by tagging them and using findobj instead. Added a script that checks a range of sensitivities around what the user inputs in an attempt to get a match more often. Also added conditions to check if we only found one circle (this is our only successful case). Added an input line for getting a real distance from the user for setting the calibration factor. Don't know when this should be called other than startup though...